@@ -1,4 +1,4 @@
-class AddNameAndAge < ActiveRecord::Migration
+class AddNameAndAge < ActiveRecord::Migration[5.0]
   def change
     add_column :users, :name, :string
     add_column :users, :age, :integer
